[KLIB tool] New command: "dump-ll-metadata" for low-level metadata dump
Unlike "dump-metadata", which produces output very similar to
the original Kotlin code, the new "dump-ll-metadata" command dumps
the metadata in tree-like structure that fully replicate the order
and all the attributes and flags exactly as it is written in protobuf.

This command is very handy as a low-level metadata disassembler and
can be used on par with the similar IR-related command "dump-ir".

^KT-62340